# Do not rely on exceptions in Struct#dig