This attempts to solve https://github.com/oracle/truffleruby/issues/2299.

Not relying on exception flow shows much better perf in interpreter mode.

```ruby
require 'benchmark/ips'

puts RUBY_DESCRIPTION

struct = Struct.new(:foo).new(14)

Benchmark.ips do |x|
  x.report('found') do
      struct[:foo]
  end

  x.report('not-found') do
    struct.dig(:bar)
  end

  x.compare!
end
```

**Before**

```
$ jt ruby demo.rb
Using Interpreted TruffleRuby: mxbuild/truffleruby-jvm
$ /src/github.com/shopify/truffleruby/mxbuild/truffleruby-jvm/languages/ruby/bin/ruby \
  --experimental-options \
  --core-load-path=/src/github.com/shopify/truffleruby/src/main/ruby/truffleruby \
  demo.rb
[ruby] WARNING gem paths: /usr/local/bundle are not marked as installed by TruffleRuby. They might belong to another Ruby implementation and break unexpectedly. Configure your Ruby manager to use TruffleRuby, or `unset GEM_HOME GEM_PATH`. See https://github.com/oracle/truffleruby/blob/master/doc/user/ruby-managers.md
truffleruby 21.1.0-dev-2855d7c9, like ruby 2.7.2, Interpreted JVM [x86_64-linux]
Warming up --------------------------------------
               found    57.241k i/100ms
           not-found     5.873k i/100ms
Calculating -------------------------------------
               found    541.688k (±13.0%) i/s -      2.690M in   5.073909s
           not-found     56.472k (±13.2%) i/s -    281.904k in   5.079920s

Comparison:
               found:   541687.5 i/s
           not-found:    56472.2 i/s - 9.59x  (± 0.00) slower
```

**After**

```
$ jt ruby demo.rb
Using Interpreted TruffleRuby: mxbuild/truffleruby-jvm
$ /src/github.com/shopify/truffleruby/mxbuild/truffleruby-jvm/languages/ruby/bin/ruby \
  --experimental-options \
  --core-load-path=/src/github.com/shopify/truffleruby/src/main/ruby/truffleruby \
  demo.rb
[ruby] WARNING gem paths: /usr/local/bundle are not marked as installed by TruffleRuby. They might belong to another Ruby implementation and break unexpectedly. Configure your Ruby manager to use TruffleRuby, or `unset GEM_HOME GEM_PATH`. See https://github.com/oracle/truffleruby/blob/master/doc/user/ruby-managers.md
truffleruby 21.1.0-dev-2855d7c9, like ruby 2.7.2, Interpreted JVM [x86_64-linux]
Warming up --------------------------------------
               found    57.851k i/100ms
           not-found    52.936k i/100ms
Calculating -------------------------------------
               found    546.247k (± 3.8%) i/s -      2.777M in   5.091750s
           not-found    536.844k (± 2.0%) i/s -      2.700M in   5.030927s

Comparison:
               found:   546247.3 i/s
           not-found:   536844.3 i/s - same-ish: difference falls within error

```