@chrisseaton 2021-03-29
Thanks for this work. 10x slower for missing is another example of a 'performance cliff'. It's hard to explain to users why a missing field is an order of magnitude worse when it seems to be conceptually the same complexity.